+++ CHANGELOG.md
@@ -5,8 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [1.0.3] - 2019.03.01
-### Fixed
+### Changed
 - Added default line-height of 18px for text elements
+- Added support for defaultTextLineHeight, defaultRowPadding, defaultColumnPadding, defaultTextPadding at main level
 
 ## [1.0.2] - 2019.02.28
 ### Fixed